Moves LostArmiesCalculator to package lostArmiesCalculator
