+++ core/src/main/combat/lostArmiesCalculator/LostArmiesCalculatorExceptions.kt
@@ -1,4 +1,4 @@
-package combat
+package combat.lostArmiesCalculator
 
 import PositiveInt
